requirements.txt
langchain
langchain-community
langchain_text_splitters
langchain-mistralai
langchain_chroma
python_dotenv